requirements.txt
feedparser
python-dotenv
anthropic
